BI OK, GC OK BDD OK

+++ IN/src/fr/epita/sigl/miwa/application/BI/CategorieVenteBI.java
@@ -3,15 +3,15 @@ package fr.epita.sigl.miwa.application.BI;
 public class CategorieVenteBI {
 	private String ref_categorie;
 	private Integer quantite_vendue;
-	private Integer montant_fournisseur;
-	private Integer montant_vente;
-	
+	private Float montant_fournisseur;
+	private Float montant_vente;
+
 	public CategorieVenteBI()
 	{
 		
 	}
 	
-	public CategorieVenteBI(String ref_categorie, Integer quantite_vendue, Integer montant_fournisseur, Integer montant_vente)
+	public CategorieVenteBI(String ref_categorie, Integer quantite_vendue, Float montant_fournisseur, Float montant_vente)
 	{
 		this.ref_categorie = ref_categorie;
 		this.quantite_vendue = quantite_vendue;
@@ -23,8 +23,8 @@ public class CategorieVenteBI {
 	{
 		this.ref_categorie = ref_categorie;
 		this.quantite_vendue = Integer.parseInt(quantite_vendue);
-		this.montant_fournisseur = Integer.parseInt(montant_fournisseur);
-		this.montant_vente = Integer.parseInt(montant_vente);
+		this.montant_fournisseur = Float.parseFloat(montant_fournisseur);
+		this.montant_vente = Float.parseFloat(montant_vente);
 	}
 	
 	public String sendXML()
@@ -52,16 +52,16 @@ public class CategorieVenteBI {
 	public void setQuantite_vendue(Integer quantite_vendue) {
 		this.quantite_vendue = quantite_vendue;
 	}
-	public Integer getMontant_fournisseur() {
+	public Float getMontant_fournisseur() {
 		return montant_fournisseur;
 	}
-	public void setMontant_fournisseur(Integer montant_fournisseur) {
+	public void setMontant_fournisseur(Float montant_fournisseur) {
 		this.montant_fournisseur = montant_fournisseur;
 	}
-	public Integer getMontant_vente() {
+	public Float getMontant_vente() {
 		return montant_vente;
 	}
-	public void setMontant_vente(Integer montant_vente) {
+	public void setMontant_vente(Float montant_vente) {
 		this.montant_vente = montant_vente;
 	}
 }
